+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 <a name="3.6.0"></a>
 # 3.6.0 (2017-10-05)
 * added drag over / leave / enter / end to action mapping
+* added $event to allowDrop function
+* Allow for node ID=0
 
 <a name="3.5.0"></a>
 # 3.5.0 (2017-29-04)